# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Velvet Store API - Premium adult wellness ecommerce for Kenya
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: tags
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: inStock
          in: query
          schema:
            type: boolean
        - name: featured
          in: query
          schema:
            type: boolean
        - name: isOffer
          in: query
          schema:
            type: boolean
        - name: isBestSeller
          in: query
          schema:
            type: boolean
        - name: isNewArrival
          in: query
          schema:
            type: boolean
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, newest, popular]
        - name: status
          in: query
          schema:
            type: string
            enum: [published, draft]
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"

    post:
      operationId: createProduct
      tags: [products, admin]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found

    put:
      operationId: updateProduct
      tags: [products, admin]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

    delete:
      operationId: deleteProduct
      tags: [products, admin]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /products/{id}/images:
    post:
      operationId: uploadProductImage
      tags: [products, admin]
      summary: Upload a product image (base64)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadImageBody"
      responses:
        "201":
          description: Uploaded image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductImage"

    delete:
      operationId: deleteProductImage
      tags: [products, admin]
      summary: Delete a product image
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: imageId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /products/{id}/related:
    get:
      operationId: getRelatedProducts
      tags: [products]
      summary: Get related products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  # ── CATEGORIES ─────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # ── PRODUCT SCHEMAS ─────────────────────────────────────────
    Product:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        sku:
          type: string
        shortDescription:
          type: string
        description:
          type: string
        price:
          type: number
        compareAtPrice:
          type: number
          nullable: true
        stockQuantity:
          type: integer
        categoryId:
          type: integer
          nullable: true
        category:
          $ref: "#/components/schemas/Category"
          nullable: true
        tags:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProductImage"
        featured:
          type: boolean
        isOffer:
          type: boolean
        isBestSeller:
          type: boolean
        isNewArrival:
          type: boolean
        status:
          type: string
          enum: [published, draft]
        seoTitle:
          type: string
          nullable: true
        seoDescription:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, slug, price, stockQuantity, featured, isOffer, isBestSeller, isNewArrival, status, createdAt, updatedAt, images, tags]
    ProductImage:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        url:
          type: string
        altText:
          type: string
          nullable: true
        sortOrder:
          type: integer
        isPrimary:
          type: boolean
      required: [id, productId, url, sortOrder, isPrimary]
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [products, total, page, totalPages]
...
    CreateProductBody:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        sku:
          type: string
        shortDescription:
          type: string
        description:
          type: string
        price:
          type: number
        compareAtPrice:
          type: number
          nullable: true
        stockQuantity:
          type: integer
        categoryId:
          type: integer
          nullable: true
        tags:
          type: array
          items:
            type: string
        featured:
          type: boolean
        isOffer:
          type: boolean
        isBestSeller:
          type: boolean
        isNewArrival:
          type: boolean
        status:
          type: string
          enum: [published, draft]
        seoTitle:
          type: string
          nullable: true
        seoDescription:
          type: string
          nullable: true
      required: [title, slug, price, stockQuantity, status]
    UpdateProductBody:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        sku:
          type: string
        shortDescription:
          type: string
        description:
          type: string
        price:
          type: number
        compareAtPrice:
          type: number
          nullable: true
        stockQuantity:
          type: integer
        categoryId:
          type: integer
          nullable: true
        tags:
          type: array
          items:
            type: string
        featured:
          type: boolean
        isOffer:
          type: boolean
        isBestSeller:
          type: boolean
        isNewArrival:
          type: boolean
        status:
          type: string
          enum: [published, draft]
        seoTitle:
          type: string
          nullable: true
        seoDescription:
          type: string
          nullable: true
    UploadImageBody:
      type: object
      properties:
        data:
          type: string
          description: base64 encoded image data
        mimeType:
          type: string
        altText:
          type: string
          nullable: true
        isPrimary:
          type: boolean
      required: [data, mimeType]

    # ── CATEGORY SCHEMAS ────────────────────────────────────────
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        productCount:
          type: integer
        createdAt:
          type: string
      required: [id, name, slug, productCount, createdAt]
...